feat: Refactor layout structure; consolidate ProfileLayout and Sidebar components, and update authentication hooks for better organization
