+++ src/hook/singleton/swrs/fetchProfile/fetchIdentityCard/useFetchSaveIDCardSwr.tsx
@@ -1,6 +1,6 @@
 import { postMutationFetcher } from "@/lib/fetcher";
 import { useContext, useState } from "react";
-import { SwrContext } from "./SwrProvider";
+import { SwrContext } from "../../SwrProvider";
 
 export interface SaveIDCardRequest {
   cardNumber: string;
