+++ src/hook/singleton/swrs/index.ts
@@ -1,8 +1,8 @@
 export * from "./SwrProvider";
-export * from "./useFetchLoginSwr";
-export * from "./useFetchLoginGoogleSwr";
-export * from "./useFetchRegisterSwr";
-export * from "./useFetchGetAllUsersSwr";
-export * from "./useFetchUploadImgSwr";
-export * from "./useFetchScanIDCardSwr";
-export * from "./useFetchSaveIDCardSwr";
+export * from "./fetchAuth/useFetchLoginSwr";
+export * from "./fetchAuth/useFetchLoginGoogleSwr";
+export * from "./fetchAuth/useFetchRegisterSwr";
+export * from "./fetchManageAccount/useFetchGetAllUsersSwr";
+export * from "./uploadImage/useFetchUploadImgSwr";
+export * from "./fetchProfile/fetchIdentityCard/useFetchScanIDCardSwr";
+export * from "./fetchProfile/fetchIdentityCard/useFetchSaveIDCardSwr";
